requirements.txt
requests
beautifulsoup4
openai
flask
anthropic
python-dotenv